Moving the scroll functionality to higher component to ensure reusability
